src/requirements.txt
django-environ==0.4.5
psycopg2-binary==2.8.4
fastapi
uvicorn
SQLAlchemy
pydantic